requirements.txt
asgiref
black
certifi
cffi
charset-normalizer
click
colorama
cryptography
defusedxml
Django
django-allauth
django-debug-toolbar
django-extensions
django-filter
django-widget-tweaks
Faker
idna
mypy-extensions
oauthlib
packaging
pathspec
platformdirs
pycparser
PyJWT
python-dateutil
python3-openid
requests
requests-oauthlib
six
sqlparse
tomli
tzdata
urllib3
mysql-connector-python
PyMySQL
mysql-connector 
django-unfold
django-import-export
django-htmx

